fix issue when using nova under a /path
