+++ package.json
@@ -9,7 +9,7 @@
         "@vue/compiler-sfc": "^3.3.4",
         "laravel-mix": "^6.0.49",
         "mix-tailwindcss": "^1.3.0",
-        "postcss": "^8.4.29",
+        "postcss": "^8.4.31",
         "tailwindcss": "^3.3.3",
         "vue-loader": "^17.2.2"
     }
